add planet.sync namespace
